[core] Move svg-icons to the internal folder
