+++ src/Avatar/Avatar.spec.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { assert } from 'chai';
+import CancelIcon from '../internal/svg-icons/Cancel';
 import { createShallow, getClasses } from '../test-utils';
-import CancelIcon from '../svg-icons/Cancel';
 import Avatar from './Avatar';
 
 describe('<Avatar />', () => {
