+++ src/Table/TablePagination.js
@@ -2,6 +2,8 @@
 
 import React from 'react';
 import PropTypes from 'prop-types';
+import KeyboardArrowLeft from '../internal/svg-icons/KeyboardArrowLeft';
+import KeyboardArrowRight from '../internal/svg-icons/KeyboardArrowRight';
 import withStyles from '../styles/withStyles';
 import IconButton from '../IconButton';
 import Input from '../Input';
@@ -10,8 +12,6 @@ import Select from '../Select';
 import TableCell from './TableCell';
 import Toolbar from '../Toolbar';
 import Typography from '../Typography';
-import KeyboardArrowLeft from '../svg-icons/KeyboardArrowLeft';
-import KeyboardArrowRight from '../svg-icons/KeyboardArrowRight';
 
 export const styles = theme => ({
   root: {
