+++ src/Table/TableSortLabel.js
@@ -3,9 +3,9 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
+import ArrowDownwardIcon from '../internal/svg-icons/ArrowDownward';
 import withStyles from '../styles/withStyles';
 import ButtonBase from '../ButtonBase';
-import ArrowDownwardIcon from '../svg-icons/ArrowDownward';
 
 export const styles = theme => ({
   root: {
