+++ src/Tabs/TabScrollButton.spec.js
@@ -1,10 +1,10 @@
 import React from 'react';
 import { assert } from 'chai';
+import KeyboardArrowLeft from '../internal/svg-icons/KeyboardArrowLeft';
+import KeyboardArrowRight from '../internal/svg-icons/KeyboardArrowRight';
 import { createShallow, createMount, getClasses } from '../test-utils';
 import TabScrollButton from './TabScrollButton';
 import ButtonBase from '../ButtonBase';
-import KeyboardArrowLeft from '../svg-icons/KeyboardArrowLeft';
-import KeyboardArrowRight from '../svg-icons/KeyboardArrowRight';
 
 describe('<TabScrollButton />', () => {
   const props = {
