+++ src/internal/svg-icons/.eslintrc.js
