+++ src/internal/svg-icons/README.md
